fix DESIGN_PATH issue

+++ csh/39_checkout_dvc_path.csh
@@ -15,6 +15,10 @@ setenv CSH_DIR $DVC_HOME/csh
 source $CSH_DIR/12_get_server.csh
 source $CSH_DIR/13_get_project.csh
 
+if ($?depth_mode == 0) then
+   set depth_mode=files
+endif
+
 # If DVC_PATH is defined and no args is specified
 # it may be called form other dvc_checkout_* command
 # this is used to preserved all option modes of parent commands 
@@ -23,21 +27,25 @@ if (($1 != "") && ($1 != ":") && ($1 != ".")) then
    setenv DVC_PATH $1
 else if ($?DVC_PATH == 0) then
    source $CSH_DIR/14_get_design.csh
-   setenv DVC_PATH $DESIGN_PHASE/$DESIGN_BLOCK/$DESIGN_STAGE/$DESIGN_VERSN
+   setenv DVC_PATH   $DESIGN_PHASE/$DESIGN_BLOCK/$DESIGN_STAGE/$DESIGN_VERSN
    setenv DESIGN_PTR $PTR_VERSN
 endif
 
 setenv DESIGN_URL  $SVN_URL/$DESIGN_PROJT/$DVC_PATH
-setenv DESIGN_PATH $PROJT_PATH/$DVC_PATH
-mkdir -p $DESIGN_PATH
 
-if ($?force_mode) then
+if (($PROJT_PATH != "") && ($PROJT_PATH != ".")) then
+   setenv DESIGN_PATH $PROJT_PATH/$DVC_PATH
+else if ($DVC_PATH == "") then
+   setenv DESIGN_PATH .
+else
+   setenv DESIGN_PATH $DVC_PATH
+endif
+
+if (($?force_mode) && ($DESIGN_PATH != "") && ($DESIGN_PATH != "."))then
    rm -f $DESIGN_PATH
 endif
 
-if ($?depth_mode == 0) then
-   set depth_mode=files
-endif
+mkdir -p $DESIGN_PATH
 
 if {(test -e $DESIGN_PATH/.svn)} then
    svn update --quiet --force $DESIGN_PATH --set-depth $depth_mode
